request change#2
- update api naming convension
