+++ src/modules/menu/dtos/get-menus.dto.ts
@@ -1,7 +1,6 @@
 import { MenuType } from '@prisma/client';
-import { Transform, Type } from 'class-transformer';
+import { Type } from 'class-transformer';
 import {
-  IsDate,
   IsEnum,
   IsInt,
   IsNotEmpty,
@@ -9,19 +8,8 @@ import {
   IsPositive,
   IsString,
 } from 'class-validator';
-import { getEndDate, getStartDate } from 'src/utils/helpers/date.helper';
 
-export class GetClaimListDto {
-  @IsNotEmpty()
-  @Transform(({ value }) => getStartDate(value))
-  @IsDate()
-  startDate: Date;
-
-  @IsNotEmpty()
-  @Transform(({ value }) => getEndDate(value))
-  @IsDate()
-  endDate: Date;
-
+export class GetMenusDto {
   @IsNotEmpty()
   @Type(() => Number)
   @IsInt()
